Add configurability

+++ src/include/com/amazonaws/kinesis/video/webrtcclient/Include.h
@@ -1232,8 +1232,8 @@ typedef struct {
                                                          //!< connection to avoid a compromised client weakening the security of the new connections.
                                                          //!<
                                                          //!< NOTE: The certificates, if specified, can be freed after the peer connection create call
-    UINT32 rollingBufferDurationInSec;
-    UINT32 rollingBufferBitrateInMBps;
+    DOUBLE rollingBufferDurationInSec;
+    DOUBLE rollingBufferBitrateInMBps;
 } RtcConfiguration, *PRtcConfiguration;
 
 /**
